Create API route to add and delete user's skills
